requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
beautifulsoup4
anthropic
python-dotenv
pyyaml
